working on splicing phonemes

+++ sound_util.py
@@ -2,4 +2,4 @@ import numpy as np
 
 
 def seconds_to_index(t: float, sr: int) -> int:
-    return np.floor(t * sr)
+    return int(np.floor(t * sr))
